cleanup

+++ miner.h
@@ -267,7 +267,7 @@ static int execute_vm(int thr_id, struct work *work, struct instance *inst, long
 
 static bool get_work(CURL *curl);
 static int work_decode(const json_t *val, struct work *work);
-static uint64_t calc_diff(uint32_t *target);
+static double calc_diff(uint32_t *target);
 static bool add_work_package(struct work_package *work_package);
 static void update_pending_cnt(uint64_t work_id, bool add);
